downloading image from static.alta3.com

+++ setup.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 echo "Setting up the cron jobs"
-rm ssh-github-mirror-cron
+FILE="ssh-github-mirror-cron"
+if test -f "$FILE"; then
+       rm ssh-github-mirror-cron
+fi
 echo "* * * * * $USER sudo docker cp $PWD/github_users.txt ssh-github-key-mirror:/home/ubuntu/github_users.txt" >> ssh-github-mirror-cron
 echo "*/10 * * * * $USER sudo docker exec ssh-github-key-mirror  /usr/bin/python3.8 /home/ubuntu/ssh-copy-id-from-github.py -j all" >> ssh-github-mirror-cron
 sudo cp ssh-github-mirror-cron /etc/cron.d/ssh-github-mirror-cron
